# ts-client/openapi-gemini.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Gemini API",
    "description": "The Gemini API allows developers to build generative AI applications using Gemini models. Gemini is our most capable model, built from the ground up to be multimodal. It can generalize and seamlessly understand, operate across, and combine different types of information including language, images, audio, video, and code. You can use the Gemini API for use cases like reasoning across text and images, content generation, dialogue agents, summarization and classification systems, and more.",
    "version": "v1beta",
    "x-google-revision": "20260410"
  },
  "servers": [
    {
      "url": "https://generativelanguage.googleapis.com",
      "description": "Global Endpoint"
    }
  ],
  "paths": {
    "/v1beta/models": {
      "parameters": [
        {
          "$ref": "#/components/parameters/alt"
        },
        {
          "$ref": "#/components/parameters/callback"
        },
        {
          "$ref": "#/components/parameters/prettyPrint"
        },
        {
          "$ref": "#/components/parameters/_.xgafv"
        }
      ],
      "get": {
        "operationId": "ListModels",
        "description": "Lists the [`Model`s](https://ai.google.dev/gemini-api/docs/models/gemini)\navailable through the Gemini API.",
        "parameters": [
          {
            "name": "pageSize",
            "in": "query",
            "description": "The maximum number of `Models` to return (per page).\n\nIf unspecified, 50 models will be returned per page.\nThis method returns at most 1000 models per page, even if you pass a larger\npage_size.",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "pageToken",
            "in": "query",
            "description": "A page token, received from a previous `ListModels` call.\n\nProvide the `page_token` returned by one request as an argument to the next\nrequest to retrieve the next page.\n\nWhen paginating, all other parameters provided to `ListModels` must match\nthe call that provided the page token.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "default": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListModelsResponse"
                }
              }
            }
          }
        },
        "security": [],
        "tags": [
          "generativelanguage"
        ]
      }
    },
    "/v1beta/models/{model}:generateContent": {
      "post": {
        "operationId": "GenerateContent",
        "description": "Generates a model response given an input `GenerateContentRequest`.\nRefer to the [text generation\nguide](https://ai.google.dev/gemini-api/docs/text-generation) for detailed\nusage information. Input capabilities differ between models, including\ntuned models. Refer to the [model\nguide](https://ai.google.dev/gemini-api/docs/models/gemini) and [tuning\nguide](https://ai.google.dev/gemini-api/docs/model-tuning) for details.",
        "parameters": [
          {
            "name": "model",
            "in": "path",
            "description": "Resource ID segment making up resource `name`. It identifies the resource within its parent collection as described in https://google.aip.dev/122.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "description": "The request body.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GenerateContentRequest"
              }
            }
          }
        },
        "responses": {
          "default": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenerateContentResponse"
                }
              }
            }
          }
        },
        "security": [],
        "tags": [
          "generativelanguage"
        ]
      },
      "parameters": [
        {
          "$ref": "#/components/parameters/alt"
        },
        {
          "$ref": "#/components/parameters/callback"
        },
        {
          "$ref": "#/components/parameters/prettyPrint"
        },
        {
          "$ref": "#/components/parameters/_.xgafv"
        }
      ]
    },
    "/v1beta/models/{model}:embedContent": {
      "parameters": [
        {
          "$ref": "#/components/parameters/alt"
        },
        {
          "$ref": "#/components/parameters/callback"
        },
        {
          "$ref": "#/components/parameters/prettyPrint"
        },
        {
          "$ref": "#/components/parameters/_.xgafv"
        }
      ],
      "post": {
        "operationId": "EmbedContent",
        "description": "Generates a text embedding vector from the input `Content` using the\nspecified [Gemini Embedding\nmodel](https://ai.google.dev/gemini-api/docs/models/gemini#text-embedding).",
        "parameters": [
          {
            "name": "model",
            "in": "path",
            "description": "Resource ID segment making up resource `name`. It identifies the resource within its parent collection as described in https://google.aip.dev/122.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "description": "The request body.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmbedContentRequest"
              }
            }
          }
        },
        "responses": {
          "default": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmbedContentResponse"
                }
              }
            }
          }
        },
        "security": [],
        "tags": [
          "generativelanguage"
        ]
      }
    },
    "/v1beta/models/{model}:streamGenerateContent": {
      "parameters": [
        {
          "$ref": "#/components/parameters/alt"
        },
        {
          "$ref": "#/components/parameters/callback"
        },
        {
          "$ref": "#/components/parameters/prettyPrint"
        },
        {
          "$ref": "#/components/parameters/_.xgafv"
        }
      ],
      "post": {
        "operationId": "StreamGenerateContent",
        "description": "Generates a [streamed\nresponse](https://ai.google.dev/gemini-api/docs/text-generation?lang=python#generate-a-text-stream)\nfrom the model given an input `GenerateContentRequest`.",
        "parameters": [
          {
            "name": "model",
            "in": "path",
            "description": "Resource ID segment making up resource `name`. It identifies the resource within its parent collection as described in https://google.aip.dev/122.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "description": "The request body.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GenerateContentRequest"
              }
            }
          }
        },
        "responses": {
          "default": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenerateContentResponse"
                }
              }
            }
          }
        },
        "security": [],
        "tags": [
          "generativelanguage"
        ]
      }
    },
...
  },
  "components": {
    "parameters": {
      "_.xgafv": {
        "description": "V1 error format.",
        "schema": {
          "type": "string",
          "enum": [
            "1",
            "2"
          ],
          "x-google-enum-descriptions": [
            "v1 error format",
            "v2 error format"
          ]
        },
        "name": "$.xgafv",
        "in": "query"
      },
      "alt": {
        "schema": {
          "type": "string",
          "default": "json",
          "enum": [
            "json",
            "media",
            "proto"
          ],
          "x-google-enum-descriptions": [
            "Responses with Content-Type of application/json",
            "Media download with context-dependent Content-Type",
            "Responses with Content-Type of application/x-protobuf"
          ]
        },
        "description": "Data format for response.",
        "name": "$alt",
        "in": "query"
      },
      "callback": {
        "name": "$callback",
        "in": "query",
        "description": "JSONP",
        "schema": {
          "type": "string"
        }
      },
      "prettyPrint": {
        "description": "Returns response with indentations and line breaks.",
        "schema": {
          "type": "boolean",
          "default": "true"
        },
        "name": "$prettyPrint",
        "in": "query"
      }
    },
    "schemas": {
...
      "EmbedContentResponse": {
        "description": "The response to an `EmbedContentRequest`.",
        "type": "object",
        "properties": {
          "embedding": {
            "description": "Output only. The embedding generated from the input content.",
            "allOf": [
              {
                "$ref": "#/components/schemas/ContentEmbedding"
              }
            ],
            "readOnly": true
          },
          "usageMetadata": {
            "description": "Output only. The usage metadata for the request.",
            "allOf": [
              {
                "$ref": "#/components/schemas/EmbeddingUsageMetadata"
              }
            ],
            "readOnly": true
          }
        }
      },
...
      "GenerateContentRequest": {
        "description": "Request to generate a completion from the model.",
        "type": "object",
        "properties": {
          "model": {
            "description": "Required. The name of the `Model` to use for generating the completion.\n\nFormat: `models/{model}`.",
            "type": "string"
          },
          "toolConfig": {
            "description": "Optional. Tool configuration for any `Tool` specified in the request. Refer to the\n[Function calling\nguide](https://ai.google.dev/gemini-api/docs/function-calling#function_calling_mode)\nfor a usage example.",
            "allOf": [
              {
                "$ref": "#/components/schemas/ToolConfig"
              }
            ]
          },
          "generationConfig": {
            "description": "Optional. Configuration options for model generation and outputs.",
            "allOf": [
              {
                "$ref": "#/components/schemas/GenerationConfig"
              }
            ]
          },
          "tools": {
            "description": "Optional. A list of `Tools` the `Model` may use to generate the next response.\n\nA `Tool` is a piece of code that enables the system to interact with\nexternal systems to perform an action, or set of actions, outside of\nknowledge and scope of the `Model`. Supported `Tool`s are `Function` and\n`code_execution`. Refer to the [Function\ncalling](https://ai.google.dev/gemini-api/docs/function-calling) and the\n[Code execution](https://ai.google.dev/gemini-api/docs/code-execution)\nguides to learn more.",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Tool"
            }
          },
          "contents": {
            "description": "Required. The content of the current conversation with the model.\n\nFor single-turn queries, this is a single instance. For multi-turn queries\nlike [chat](https://ai.google.dev/gemini-api/docs/text-generation#chat),\nthis is a repeated field that contains the conversation history and the\nlatest request.",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Content"
            }
          },
          "cachedContent": {
            "description": "Optional. The name of the content\n[cached](https://ai.google.dev/gemini-api/docs/caching) to use as context\nto serve the prediction. Format: `cachedContents/{cachedContent}`",
            "type": "string"
          },
          "safetySettings": {
            "description": "Optional. A list of unique `SafetySetting` instances for blocking unsafe content.\n\nThis will be enforced on the `GenerateContentRequest.contents` and\n`GenerateContentResponse.candidates`. There should not be more than one\nsetting for each `SafetyCategory` type. The API will block any contents and\nresponses that fail to meet the thresholds set by these settings. This list\noverrides the default settings for each `SafetyCategory` specified in the\nsafety_settings. If there is no `SafetySetting` for a given\n`SafetyCategory` provided in the list, the API will use the default safety\nsetting for that category. Harm categories HARM_CATEGORY_HATE_SPEECH,\nHARM_CATEGORY_SEXUALLY_EXPLICIT, HARM_CATEGORY_DANGEROUS_CONTENT,\nHARM_CATEGORY_HARASSMENT, HARM_CATEGORY_CIVIC_INTEGRITY are supported.\nRefer to the [guide](https://ai.google.dev/gemini-api/docs/safety-settings)\nfor detailed information on available safety settings. Also refer to the\n[Safety guidance](https://ai.google.dev/gemini-api/docs/safety-guidance) to\nlearn how to incorporate safety considerations in your AI applications.",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SafetySetting"
            }
          },
          "serviceTier": {
            "description": "Optional. The service tier of the request.",
            "allOf": [
              {
                "$ref": "#/components/schemas/ServiceTier"
              }
            ]
          },
          "systemInstruction": {
            "description": "Optional. Developer set [system\ninstruction(s)](https://ai.google.dev/gemini-api/docs/system-instructions).\nCurrently, text only.",
            "allOf": [
              {
                "$ref": "#/components/schemas/Content"
              }
            ]
          },
          "store": {
            "description": "Optional. Configures the logging behavior for a given request. If set, it takes\nprecedence over the project-level logging config.",
            "type": "boolean"
          }
        },
        "required": [
          "model",
          "contents"
        ]
      },
...
      "ServiceTier": {
        "type": "string",
        "enum": [
          "unspecified",
          "standard",
          "flex",
          "priority"
        ],
        "x-google-enum-descriptions": [
          "Default service tier, which is standard.",
          "Standard service tier.",
          "Flex service tier.",
          "Priority service tier."
        ]
      },
...
      "ListModelsResponse": {
        "description": "Response from `ListModel` containing a paginated list of Models.",
        "type": "object",
        "properties": {
          "nextPageToken": {
            "description": "A token, which can be sent as `page_token` to retrieve the next page.\n\nIf this field is omitted, there are no more pages.",
            "type": "string"
          },
          "models": {
            "description": "The returned Models.",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Model"
            }
          }
        }
      },
...
      "GenerateContentResponse": {
        "description": "Response from the model supporting multiple candidate responses.\n\nSafety ratings and content filtering are reported for both\nprompt in `GenerateContentResponse.prompt_feedback` and for each candidate\nin `finish_reason` and in `safety_ratings`. The API:\n - Returns either all requested candidates or none of them\n - Returns no candidates at all only if there was something wrong with the\n   prompt (check `prompt_feedback`)\n - Reports feedback on each candidate in `finish_reason` and\n   `safety_ratings`.",
        "type": "object",
        "properties": {
          "candidates": {
            "description": "Candidate responses from the model.",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Candidate"
            }
          },
          "modelVersion": {
            "description": "Output only. The model version used to generate the response.",
            "type": "string",
            "readOnly": true
          },
          "usageMetadata": {
            "description": "Output only. Metadata on the generation requests' token usage.",
            "allOf": [
              {
                "$ref": "#/components/schemas/UsageMetadata"
              }
            ],
            "readOnly": true
          },
          "promptFeedback": {
            "description": "Returns the prompt's feedback related to the content filters.",
            "allOf": [
              {
                "$ref": "#/components/schemas/PromptFeedback"
              }
            ]
          },
          "modelStatus": {
            "description": "Output only. The current model status of this model.",
            "allOf": [
              {
                "$ref": "#/components/schemas/ModelStatus"
              }
            ],
            "readOnly": true
          },
          "responseId": {
            "description": "Output only. response_id is used to identify each response.",
            "type": "string",
            "readOnly": true
          }
        }
      },
...
      "EmbedContentRequest": {
        "description": "Request containing the `Content` for the model to embed.",
        "type": "object",
        "properties": {
          "content": {
            "allOf": [
              {
                "$ref": "#/components/schemas/Content"
              }
            ],
            "description": "Required. The content to embed. Only the `parts.text` fields will be counted."
          },
          "taskType": {
            "description": "Optional. Optional task type for which the embeddings will be used. Not supported on\nearlier models (`models/embedding-001`).",
            "allOf": [
              {
                "$ref": "#/components/schemas/TaskType"
              }
            ]
          },
          "title": {
            "description": "Optional. An optional title for the text. Only applicable when TaskType is\n`RETRIEVAL_DOCUMENT`.\n\nNote: Specifying a `title` for `RETRIEVAL_DOCUMENT` provides better quality\nembeddings for retrieval.",
            "type": "string"
          },
          "model": {
            "description": "Required. The model's resource name. This serves as an ID for the Model to use.\n\nThis name should match a model name returned by the `ListModels` method.\n\nFormat: `models/{model}`",
            "type": "string"
          },
          "outputDimensionality": {
            "description": "Optional. Optional reduced dimension for the output embedding. If set, excessive\nvalues in the output embedding are truncated from the end. Supported by\nnewer models since 2024 only. You cannot set this value if using the\nearlier model (`models/embedding-001`).",
            "type": "integer",
            "format": "int32"
          }
        },
        "required": [
          "model",
          "content"
        ]
      },
...
    },
    "securitySchemes": {
      "google_oauth_implicit": {
        "type": "oauth2",
        "description": "Google Oauth 2.0 implicit authentication flow.",
        "flows": {
          "implicit": {
            "authorizationUrl": "https://accounts.google.com/o/oauth2/v2/auth",
            "scopes": {
              "https://www.googleapis.com/auth/devstorage.read_only": "View your data in Google Cloud Storage"
            }
          }
        }
      },
      "google_oauth_code": {
        "type": "oauth2",
        "description": "Google Oauth 2.0 authorizationCode authentication flow.",
        "flows": {
          "authorizationCode": {
            "authorizationUrl": "https://accounts.google.com/o/oauth2/v2/auth",
            "tokenUrl": "https://oauth2.googleapis.com/token",
            "refreshUrl": "https://oauth2.googleapis.com/token",
            "scopes": {
              "https://www.googleapis.com/auth/devstorage.read_only": "View your data in Google Cloud Storage"
            }
          }
        }
      },
      "bearer_auth": {
        "type": "http",
        "description": "Http bearer authentication.",
        "scheme": "bearer"
      }
    }
  },
...
}